CREATE strict server

+++ internal/models/post.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Post struct {
-	Id       int64
-	UserId   int64
+	Id       int
+	UserId   int
 	Text     string
 	Datetime time.Time
 }
